==> R/01_Initialization.Rmd <==
---
title: "01_Initialization"
author: "Marc Leman"
date: "14/11/2023"
output: html_document
---

# Set up R environment
# The following code sets up options for the R environment in the document.
# 'echo = TRUE' specifies that R code in subsequent chunks will be displayed.
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Set your working directory:
# This is the directory where we begin with knitr.
# The following code clears the workspace at the beginning of the document.
# 'rm(list = ls())' removes all objects in the global environment.
```{r initialization, echo=TRUE}
rm(list = ls())
```
  
# Load libraries
# The following code loads various R packages used in the analysis.
# Uncomment the 'install.packages' lines to install the packages if not already installed.
```{r download packages if needed, echo=TRUE}
library(tidyverse)
#library(reshape2)
library(readxl)
library(readr)
library(ggplot2)
library(patchwork)
library(grid)
library(gridExtra)
library(ggExtra)
library(ggthemes)
library(ggside)
#library(ggpubr)
#library(sjPlot)
#install.packages("coda")
#install.packages("rjags") -> vergt eerst aparte installatie
#install.packages("BEST")
library(coda)
#library(rjags)
#library(BEST)
library(rstan)
options(mc.cores = parallel::detectCores())
rstan_options(auto_write = TRUE)
library(tidybayes)
library(bayesplot)
options(mc.cores = parallel::detectCores())
library(brms)
#library(shinystan)
library(bayestestR)
library(posterior)
#library(lme4)
#library(BListener)
#library(flexdashboard)
##library(sjPlot)
#library(kableExtra)
#library(magick)
#library(modelbased)
#library(emmeans)
#library(patchwork)
#install.packages("easystats", repos = "https://easystats.r-universe.dev")
#library("easystats")
#library(insight)
#library(see)
#library(equatiomatic)
library(performance)
library(parameters)
#library(see)
library(mgcv)
#library(smoothr)
library(gratia)
library(foreach)
library(doParallel)
#library(DescTools)
#install.packages("ggcorrplot")                     
library("ggcorrplot") 
#remotes::install_github("stan-dev/cmdstanr")
library(cmdstanr)
check_cmdstan_toolchain(fix = TRUE, quiet = TRUE)
check_cmdstan_toolchain()
#install_cmdstan(cores = 16, overwrite=TRUE)
cmdstan_path()
library(xfun)
```
