==> R/02_DataPreparation.Rmd <==
---
title: "02_DataPreparation"
author: "Marc Leman"
date: "14/11/2023"
output:
  pdf_document:
    latex_engine: xelatex
  html_document: default
  word_document: default
---

# Set up R environment
```{r setup, include=FALSE}
# Set up R environment options.
knitr::opts_chunk$set(echo = TRUE)
# Custom chunk hook to handle font size in LaTeX output.
def.chunk.hook  <- knitr::knit_hooks$get("chunk")
knitr::knit_hooks$set(chunk = function(x, options) {
  x <- def.chunk.hook(x, options)
  ifelse(options$size != "normalsize", paste0("\n \\", options$size,"\n\n", x, "\n\n \\normalsize"), x)
})
```


### Step 1: Construct Data
```{r getAgeMSE, echo=TRUE}
# Read the CSV file and preprocess data for analysis.
TotalDataSetTidy <- read_csv("DataArray.csv", col_names = FALSE)
Data2 <- TotalDataSetTidy %>% mutate(response = (X1), 
                                         violin = factor(X2), 
                                         piece = factor(X3),
                                         condition = factor(X4), 
                                         trial = factor(X5), 
                                         session = as.integer(X5),
                                         participant = factor(X6),
                                         data_index = as.integer(X7), 
                                         timestep = data_index,
                                         data_type = factor(as.integer(X8), labels = c("procustus","sparc","bowlength","WPQ","MPQS","MPQP","Difficulty")) 
)

# Transform the data and create separate datasets for different types.
procustus <- Data2 %>% filter(data_type == "procustus") %>% 
  mutate(responseScaled = scale(response, center = TRUE, scale = TRUE),
         responseN_procustus = responseScaled - min(responseScaled)
  ) %>%
  select(violin, piece, condition, trial, session, participant, timestep, data_type, responseN_procustus)

sparc <- Data2 %>% filter(data_type == "sparc") %>% 
  mutate(responseScaled = scale(response, center = TRUE, scale = TRUE),
         responseN_sparc = responseScaled - min(responseScaled),
         logresponseN_sparc = log(1 + responseN_sparc)) %>%
  select(responseN_sparc)

bowlength <- Data2 %>% filter(data_type == "bowlength") %>%
  mutate(responseScaled = scale(response, center = TRUE, scale = TRUE),
         responseN_bowlength = responseScaled - min(responseScaled),
         responseN_bowlength = log(1 + responseN_bowlength)) %>%
  select(responseN_bowlength)

# Questionnaires
WPQ <- Data2 %>% filter(data_type == "WPQ") %>%mutate(WPQ = response) %>% select(participant,condition,trial,WPQ)
MPQS <- Data2 %>% filter(data_type == "MPQS") %>%mutate(MPQS = response) %>% select(MPQS)
MPQP <- Data2 %>% filter(data_type == "MPQP") %>%mutate(MPQP = response) %>% select(MPQP)
Difficulty <- Data2 %>% filter(data_type == "Difficulty") %>% mutate(Difficulty = response,Difficulty = scale(Difficulty)) %>% select(Difficulty) 

data_questionnaires <- cbind(WPQ,MPQS,MPQP,Difficulty) 
data_metric <- cbind(procustus,sparc,bowlength)
Data <- data_metric

rm(df.extended,df.extended.11,df.extended.12,df.extended.21,df.extended.22,
   df.11,df.12,df.21,df.22, 
   Legend, Legende, TotalDataSetTidy,
   Data2, 
   procustus, sparc, bowlength,
   WPQ,MPQS,MPQP,Difficulty
   ) 

# Visualization and exploration of the data.
min(data_metric$responseN_procustus)
max(data_metric$responseN_procustus)
hist(data_metric$responseN_procustus,n=100)

p001 <- ggplot(data_metric) +
  geom_point(aes(x=participant, y = responseN_procustus, color = condition), 
             position = position_jitterdodge(jitter.width = .5, dodge.width = .6), size = .4, alpha = .4) +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=1)) +
    facet_grid(~ piece)

p002 <- ggplot(data_metric) +
  geom_point(aes(x=timestep, y = responseN_procustus, color = trial), 
             position = position_jitterdodge(jitter.width = .2,dodge.width = .6)) +
  facet_grid(~ piece + condition) +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=1))

hist(data_metric$responseN_procustus,n=100)
p01 <- ggplot(data_metric) +
  geom_point(aes(x=timestep, y = responseN_procustus, color = trial), 
             position = position_jitterdodge(jitter.width = .2,dodge.width = .6)) +
  facet_grid(~ piece + violin + condition) +
  ggtitle("Procustus") +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=1))

hist(data_metric$responseN_sparc,n=100)
p02 <- ggplot(data_metric) +
  geom_point(aes(x=timestep, y = responseN_sparc, color = trial), 
             position = position_jitterdodge(jitter.width = .2,dodge.width = .6)) +
  facet_grid(~ piece + violin + condition) +
  ggtitle("Sparc") +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust=1))

# m0 is median of the group
m0 <- data_metric %>%  group_by(piece,violin) %>% 
  summarize(responseGroup_procustus = median(responseN_procustus),
            responseGroup_sparc = median(responseN_sparc)
  )
dummy <- merge(data_metric,m0)

# m1 has response0M (non-calibrated) and response1M (calibrated) per part. cond.
m1 <-  dummy %>% 
  group_by(participant, condition, trial) %>%  
  summarize(response1M_procustus = median(responseN_procustus - responseGroup_procustus), # calibrated
            response0M_procustus = median(responseN_procustus),
            response1M_sparc = median(responseN_sparc - responseGroup_sparc), # calibrated
            response0M_sparc = median(responseN_sparc)
  ) 

# Merge with data_participants and data_questionnaires
# # new Data variable!
Data <- merge(m1,data_questionnaires)
Data <- Data %>% mutate(log_response0M_procustus = log(response0M_procustus),
                        log_response1M_procustus = log(response1M_procustus),
                        log_response0M_sparc = log(response0M_sparc),
                        log_response1M_sparc = log(response1M_sparc),
                        difficulty_s = scale(Difficulty))

rm(dummy,m0,m1)

p1 <- ggplot(Data) +
  geom_point(aes(x=participant,y=response0M_procustus, color = condition), position = position_jitterdodge()) +
  ggtitle("procustus")
p2 <- ggplot(Data) +
  geom_point(aes(x=participant,y=response1M_procustus, color = condition), position = position_jitterdodge()) +
  ggtitle("procustus-calibrated")
p3 <- ggplot(Data) +
  geom_point(aes(x=participant,y=response0M_sparc, color = condition), position = position_jitterdodge()) +
  ggtitle("sparc")
p4 <- ggplot(Data) +
  geom_point(aes(x=participant,y=response1M_sparc, color = condition), position = position_jitterdodge())+
  ggtitle("sparc-calibrated")
p5 <- ggplot(Data) +
  geom_point(aes(x=response0M_procustus,y=scale(Difficulty), color = condition), position = position_jitterdodge())
```

# Create summary plots for Procustus and Sparc.
```{r getAgeMSE, echo=TRUE}
p001
p002
p01
p02
```

### Step 2: Save Data
```{r storedata, echo=TRUE}
# Clean the data and save it for further analysis.
Data <- Data %>% mutate(condition = factor(condition),
                                                     participant = factor(participant),
                                                     trial = factor(trial)
                                                     )

Data <- Data %>% mutate_all(~ifelse(is.nan(.), NA, .)) %>% mutate(condition = factor(condition),
                                                     participant = factor(participant),
                                                     trial = factor(trial)
                                                     )

# Handle missing values in the data.
w <- which(is.na(Data$Difficulty))
Data$Difficulty[w] = 0
w <- which(is.na(Data$difficulty_s))
Data$difficulty_s[w] = 0

# Save the cleaned data to a file.
save(file = "Data.RData", Data)
```
