==> R/03_Functions.Rmd <==
---
title: "03_Functions"
author: "Marc Leman"
date: "14/11/2023"
output: html_document
---

# Set up R environment
```{r setup, include=FALSE}
# Set up R environment options.
knitr::opts_chunk$set(echo = TRUE)
```

# Modelling functions GUSO paper
```{r modellingfunctions, echo=TRUE}
# Load cleaned data
load(file = "Data.RData")

# Function to run the Bayesian model using cmdstanr
run_model_cmdstanr <- function(Data, form, fam, priors){
  print(paste(form[1],fam,sep="\n"))
  fit <- brm(data = Data,
             formula = form,
             family = fam,
             prior = priors,
             iter = 2000, warmup = 500, # for testing, otherwise, use 4000 and 1000
            # control = list(adapt_delta = 0.995,  max_treedepth = 12),
             # init = 0,
             # thin = 2,
             chains = 2,
             backend = "cmdstanr",
             threads = threading(2),
             save_pars = save_pars(all = TRUE),
             #sample_prior = TRUE #,
             #file = filen
  )
  return(fit)
}

# Function to perform Bayesian model comparison
do_BF_comparison <- function(fit1,fit2){
  bf <- bayes_factor(fit1,fit2,log=TRUE)
  return(bf)
}

# Function for post-analysis of the Bayesian model
do_post_analysis <- function(fit){
  # fit <- model_procustus0M
  library(patchwork)
  t <- fit$formula
  print(t)
  p <- pp_check(fit, ndraws = 50)
  # Visualization and summary of the Bayesian model results.
  m <- mcmc_plot(fit, regex = TRUE, variable = "^[bs]_")
  #ce <- plot(conditional_effects(fit, re_formula = NA), ask = FALSE)[[1]]
  #sd <- fit %>% spread_draws(`r_condition:trial`[A,Intercept]) %>% 
  #  ggplot(aes(`r_condition:trial`,A)) + stat_halfeye()
  p01 <- (p + m) #/ (ce + sd)
  print(p01)
  #s <- summary(fit) 
  b <- r2_bayes(fit) %>% 
    data.frame() %>% 
    mutate(across(where(is.numeric), round, 2))
  print(b)
  print("calculating parameters... can take some time")
  par <- parameters(fit, test = c("pd","bf"),effects = "all",centrality = "mean") %>% 
    data.frame() %>% 
    mutate(across(where(is.numeric), round, 2))
  print(par)
#  return(list(p,m,ce,sd,b,par))
  return(list(t,p01,b,par))
}

# Function to perform hypothesis testing
do_hypothesis_test <- function(fit){
  # Contrast condition2 condition1
  h1 <- hypothesis(fit,c("b_condition1 < b_condition2") ,class = NULL)$hypothesis
  t1 <- "c12"
  # contrast trials
  #Condition1
  h2 <- hypothesis(fit,c("r_condition:trial[1_1,Intercept] > r_condition:trial[1_2,Intercept]") ,class = NULL)$hypothesis
  t2 <- "c1t12"
  h3 <- hypothesis(fit,c("r_condition:trial[1_1,Intercept] > r_condition:trial[1_3,Intercept]") ,class = NULL)$hypothesis
  t3 <- "c1t13"
  h4 <- hypothesis(fit,c("r_condition:trial[1_1,Intercept] > r_condition:trial[1_4,Intercept]") ,class = NULL)$hypothesis
    t4 <- "c1t14"
  h5 <- hypothesis(fit,c("r_condition:trial[1_2,Intercept] > r_condition:trial[1_3,Intercept]") ,class = NULL)$hypothesis
    t5 <- "c1t23"
  h6 <- hypothesis(fit,c("r_condition:trial[1_2,Intercept] > r_condition:trial[1_4,Intercept]") ,class = NULL)$hypothesis
    t6 <- "c1t24"
  h7 <- hypothesis(fit,c("r_condition:trial[1_3,Intercept] > r_condition:trial[1_4,Intercept]") ,class = NULL)$hypothesis
    t7 <- "c1t34"
    
  # Condition2
  h8 <- hypothesis(fit,c("r_condition:trial[2_1,Intercept] > r_condition:trial[2_2,Intercept]") ,class = NULL)$hypothesis
    t8 <- "c2t12"
  h9 <- hypothesis(fit,c("r_condition:trial[2_1,Intercept] > r_condition:trial[2_3,Intercept]") ,class = NULL)$hypothesis
    t9 <- "c2t13"
  h10 <- hypothesis(fit,c("r_condition:trial[2_1,Intercept] > r_condition:trial[2_4,Intercept]") ,class = NULL)$hypothesis
  t10 <- "c2t14"
  h11 <- hypothesis(fit,c("r_condition:trial[2_2,Intercept] > r_condition:trial[2_3,Intercept]") ,class = NULL)$hypothesis
  t11 <- "c2t23"
  h12 <- hypothesis(fit,c("r_condition:trial[2_2,Intercept] > r_condition:trial[2_4,Intercept]") ,class = NULL)$hypothesis
  t12 <- "c2t24"
  h13 <- hypothesis(fit,c("r_condition:trial[2_3,Intercept] > r_condition:trial[2_4,Intercept]") ,class = NULL)$hypothesis
  t13 <- "c2t34"
  
  # Condition12
  h14 <- hypothesis(fit,c("b_condition1 + r_condition:trial[1_1,Intercept] > b_condition2 + r_condition:trial[2_1,Intercept]") ,class = NULL)$hypothesis
  t14 <- "c12t1"
  h15 <- hypothesis(fit,c("b_condition1 + r_condition:trial[1_2,Intercept] > b_condition2 + r_condition:trial[2_2,Intercept]") ,class = NULL)$hypothesis
  t15 <- "c12t2"
  h16 <- hypothesis(fit,c("b_condition1 + r_condition:trial[1_3,Intercept] > b_condition2 + r_condition:trial[2_3,Intercept]") ,class = NULL)$hypothesis
  t16 <- "c12t3"
  h17 <- hypothesis(fit,c("b_condition1 + r_condition:trial[1_4,Intercept] > b_condition2 + r_condition:trial[2_4,Intercept]") ,class = NULL)$hypothesis
  t17 <- "c12t4"
    
  h <- rbind(h1,h2,h3,h4,h5,h6,h7,h8,h9,h10,h11,h12,h13,h14,h15,h16,h17) %>% data.frame() %>% mutate(across(where(is.numeric), round, 2))
  t <- rbind(t1,t2,t3,t4,t5,t6,t7,t8,t9,t10,t11,t12,t13,t14,t15,t16,t17) %>% data.frame()
  names(t) <- "Label"
  t <- t %>% mutate(Label = factor(Label,labels = Label, levels = Label) )
  
  h <- cbind(h,t) %>% mutate(Tag =  paste(Estimate,"[",CI.Lower,",", CI.Upper, "] --", Post.Prob, Star ) ) 
  
  # to print
  hh <- h %>% select(Hypothesis,Label,Estimate,CI.Lower,CI.Upper,Post.Prob,Star)
  
  p1 <- h %>% ggplot() +
  geom_errorbarh(aes(xmin=CI.Lower,xmax=CI.Upper,y=Label) ) +
  geom_point(aes(x=Estimate,y=Label)) +
#  geom_text(aes(x=.85,label=(tag), y=rowname),size=3) +
  #xlim(-.25,1) +
  theme_bw() +
  ylab("Contrasts") +
  xlab("Posterior distribution (median, ci-95%)")

  p2 <- h %>% ggplot()+
 # geom_errorbarh(aes(xmin=lowerCI95,xmax=upperCI95,y=rowname) ) +
  # geom_point(aes(x=median,y=rowname)) +
  geom_text(aes(x=.85,label=Tag, y=Label),size=3) +
  #xlim(-.25,1) +
  theme_void() +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.title.y=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        )
p <- (p1 + p2)
  
  return(list(hh,p))
}

# Function for model comparison
do_comparison <- function(fit1,fit2){
  bf <- bayes_factor(fit1,fit2)
  l <- loo(fit1,fit2)
  return(list(bf,l))
}

# Function for multi-model post-analysis
do_post_analysis_Multi <- function(fit) {
  t <- fit$formula
    m <- mcmc_plot(fit, regex = TRUE, variable = "^b_")
  #ce <- plot(conditional_effects(fit, re_formula = NA), ask = FALSE)[[1]]
  #  p1 <- pp_check(fit, ndraws = 50,resp = "WPQ")
  #  p2 <- pp_check(fit, ndraws = 50,resp = "MPQS")
  #  p3  <- pp_check(fit, ndraws = 50,resp = "MPQP")

  # sd1 <- fit %>% spread_draws(`r_condition:trial__WPQ`[A,Intercept]) %>% 
  #    ggplot(aes(`r_condition:trial__WPQ`,A)) + stat_halfeye()
  # sd2 <- fit %>% spread_draws(`r_condition:trial__MPQS`[A,Intercept]) %>% 
  #   ggplot(aes(`r_condition:trial__MPQS`,A)) + stat_halfeye()
  # sd3 <- fit %>% spread_draws(`r_condition:trial__MPQP`[A,Intercept]) %>% 
  #   ggplot(aes(`r_condition:trial__MPQP`,A)) + stat_halfeye()
  # 
  #p01 <- (m + p1) / (p2 + p3)
  #s <- summary(fit) 
  bb <- r2_bayes(fit)
  b<- head(bb) %>% data.frame %>%  mutate(across(where(is.numeric), round, 2))
  #print(b)
  par <- parameters(fit, test = c("pd"), effects = "all",centrality = "mean") %>% 
    data.frame() %>% 
    mutate(across(where(is.numeric), round, 2))
return = list(m, b, par)
}
```

# General functions for plotting
```{r echo=TRUE}
# Functions for generating different types of figures.
# Save figure as a PDF file.
make_pdf_figure <- function(fn,fig){
  fn_pdf = paste(fn,".pdf",sep="")
  print(paste("plotting", fn_pdf))
  pdf(file = fn_pdf ,width=9, height=8 )  #Note that you can convert inches to centimeters dividing by 2.54.
  print(fig)
  dev.off()
}

# Save figure as a PNG file.
make_png_figure <- function(fn,fig){
  fn_png = paste(fn,".png",sep="")
  print(paste("plotting", fn_png))
  png(filename=fn_png, width=7, height=3, units = "in", res = 300)
  print(fig)
  dev.off()
}

# Save figure as a SVG file.
make_svg_figure <- function(fn,fig){
  fn_svg = paste(fn,".svg",sep="")
  print(paste("plotting", fn_svg))
  svg(file = fn_svg ,width=7, height=3
      #Note that you can convert inches to centimeters dividing by 2.54.
  ) 
  print(fig)
  dev.off()
}

# Save figure as a EPS file.
make_eps_figure <- function(fn,fig){
  fn_eps = paste(fn,".eps",sep="")
  print(paste("plotting", fn_eps))
  postscript(file = fn_eps ,width=7, height=3
             #Note that you can convert inches to centimeters dividing by 2.54.
  ) 
  print(fig)
  dev.off()
}

# Generate and save figures in multiple formats.
make_fig <- function(fn,fig){
  make_pdf_figure(fn,fig)
  make_png_figure(fn,fig)
  make_svg_figure(fn,fig)
  make_eps_figure(fn,fig)
}
```

# Function to calculate contrasts
Used for calculating contrasts of conditions in procustus, sparc
```{r echo=TRUE}
# Calculate contrasts and generate plots.
do_contrast <- function(fit){

#fit <- model_procustus0M_d2
#fit <- model_procustus1M_d2

s <- summary(fit)
c <- d2_coef <- coef(fit) 
#d2_ranef <- ranef(fit) 

cc <- coef(fit,summary = FALSE) %>% as.data.frame() %>% 
  mutate(
  condition.1.Intercept = condition.trial.1_1.condition1,
  condition.2.Intercept = condition.trial.1_1.condition2,

  diff_c12 = condition.1.Intercept - condition.2.Intercept, 
  diff_c12_t1 = condition.1.Intercept + condition.trial.1_1.Intercept - (condition.2.Intercept + condition.trial.2_1.Intercept),
  diff_c12_t2 = condition.1.Intercept + condition.trial.1_2.Intercept - (condition.2.Intercept + condition.trial.2_2.Intercept),
  diff_c12_t3 = condition.1.Intercept + condition.trial.1_3.Intercept - (condition.2.Intercept + condition.trial.2_3.Intercept),
  diff_c12_t4 = condition.1.Intercept + condition.trial.1_4.Intercept - (condition.2.Intercept + condition.trial.2_4.Intercept),
  
  diff_c1_t12 = condition.1.Intercept + condition.trial.1_1.Intercept - (condition.1.Intercept + condition.trial.1_2.Intercept),
  diff_c1_t13 = condition.1.Intercept + condition.trial.1_1.Intercept - (condition.1.Intercept + condition.trial.1_3.Intercept),
  diff_c1_t14 = condition.1.Intercept + condition.trial.1_1.Intercept - (condition.1.Intercept + condition.trial.1_4.Intercept),
  diff_c1_t23 = condition.1.Intercept + condition.trial.1_2.Intercept - (condition.1.Intercept + condition.trial.1_3.Intercept),
  diff_c1_t24 = condition.1.Intercept + condition.trial.1_2.Intercept - (condition.1.Intercept + condition.trial.1_4.Intercept),
  diff_c1_t34 = condition.1.Intercept + condition.trial.1_3.Intercept - (condition.1.Intercept + condition.trial.1_4.Intercept),

  diff_c2_t12 = condition.2.Intercept + condition.trial.2_1.Intercept - (condition.2.Intercept + condition.trial.2_2.Intercept),
  diff_c2_t13 = condition.2.Intercept + condition.trial.2_1.Intercept - (condition.2.Intercept + condition.trial.2_3.Intercept),
  diff_c2_t14 = condition.2.Intercept + condition.trial.2_1.Intercept - (condition.2.Intercept + condition.trial.2_4.Intercept),
  diff_c2_t23 = condition.2.Intercept + condition.trial.2_2.Intercept - (condition.2.Intercept + condition.trial.2_3.Intercept),
  diff_c2_t24 = condition.2.Intercept + condition.trial.2_2.Intercept - (condition.2.Intercept + condition.trial.2_4.Intercept),
  diff_c2_t34 = condition.2.Intercept + condition.trial.2_3.Intercept - (condition.2.Intercept + condition.trial.2_4.Intercept)
  )   %>% 
  select(starts_with("d")) 

# median_hdi takes median, lowerci and upperci for each column
cc_summary <- cc %>% summarise( across(everything(), median_hdi, .width=.95 ) )

# get probability of direction
cc_pd <- pd(cc)

# set cc properly and create tags in DD
CC <- data.frame()
for(i in c(1:length(cc))){
CC <- rbind(CC,(round(cc_summary[[i]][1:3],2) ) )
}
rownames(CC) <- colnames(cc_summary)
colnames(CC) <- c("median", "lowerCI95", "upperCI95")
tag <- CC$median %>% as.character()
DD <- CC %>% rownames_to_column() %>% 
  mutate(rowname = factor(rowname,colnames(cc_summary)),
         tag =  paste(median,"[",lowerCI95,",", upperCI95, "] --", round(cc_pd$pd,3)) )

# Errorbars
p2 <- DD %>% ggplot()+
  geom_errorbarh(aes(xmin=lowerCI95,xmax=upperCI95,y=rowname) ) +
  geom_point(aes(x=median,y=rowname)) +
#  geom_text(aes(x=.85,label=(tag), y=rowname),size=3) +
  #xlim(-.25,1) +
  theme_bw() +
  ylab("Contrasts") +
  xlab("Posterior distribution (median, ci-95%)")

# Table
p3 <- DD %>% ggplot()+
 # geom_errorbarh(aes(xmin=lowerCI95,xmax=upperCI95,y=rowname) ) +
  # geom_point(aes(x=median,y=rowname)) +
  geom_text(aes(x=.85,label=(tag), y=rowname),size=3) +
  #xlim(-.25,1) +
  theme_void() +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.title.y=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        )
p <- (p2 + p3)

return(list(s,p))
}
```
