==> R/04_Modelling.Rmd <==
---
title: "04_Modelling"
author: "Marc Leman"
date: "14/11/2023"
output: html_document
---

# Set up R environment
```{r setup, include=FALSE}
# Set up R environment options.
knitr::opts_chunk$set(echo = TRUE)
```

# Modelling non-calibrated procustus and sparc and save
```{r modelling, echo=TRUE}
# Load cleaned data
load(file = "Data.RData")

# non-calibrated models 0M

# Define non-calibrated models 0M
form0M_procustus <- bf(response0M_procustus ~ 0 + condition  +  (1 | condition:participant) + (1|condition:trial) )  
form0M_sparc <- bf(response0M_sparc         ~ 0 + condition  +  (1 | condition:participant) + (1|condition:trial) )  
form0M_procustus_d <- bf(response0M_procustus ~ 0 + condition  * difficulty_s + (1 + difficulty_s | condition:participant) + (1|condition:trial) )  
form0M_sparc_d <- bf(response0M_sparc         ~ 0 + condition  *  difficulty_s + (1 + difficulty_s| condition:participant) + (1|condition:trial) )  

# Define priors
#prior0M <- get_prior( form0M_procustus , data = Data)
prior0M <- c(brms::prior(normal(0.5, .25), class = "b"))

# Run models
model_procustus0M <- run_model_cmdstanr(Data,form0M_procustus,"skew_normal",prior0M)
prior0M <- c(brms::prior(normal(10, .25), class = "b"))
model_sparc0M <- run_model_cmdstanr(Data,form0M_sparc,"skew_normal",prior0M)
prior0M_d <- c(brms::prior(normal(0.5, .25), class = "b"))
model_procustus0M_d <- run_model_cmdstanr(Data,form0M_procustus_d,"skew_normal",prior0M_d)
prior0M_d <- c(brms::prior(normal(10, .25), class = "b"))
model_sparc0M_d <- run_model_cmdstanr(Data,form0M_sparc_d,"skew_normal",prior0M_d)

# Save models
save(file = "Fit/model_procustus0M.RData", model_procustus0M)
save(file = "Fit/model_sparc0M.RData", model_sparc0M)
save(file = "Fit/model_procustus0M_d.RData", model_procustus0M_d)
save(file = "Fit/model_sparc0M_d.RData", model_sparc0M_d)
```

# Modelling calibrated procustus and sparc and save
```{r modelling, echo=TRUE}
# Load data
load(file = "Data.RData")

# Define calibrated models 1M
form1M_procustus <- bf(response1M_procustus ~ 0 + condition  +  (1 | condition:participant) + (1|condition:trial) )  
form1M_sparc <- bf(response1M_sparc         ~ 0 + condition  +  (1 | condition:participant) + (1|condition:trial) )  
form1M_procustus_d <- bf(response1M_procustus ~ 0 + condition  * difficulty_s + (1 + difficulty_s | condition:participant) + (1|condition:trial) )  
form1M_sparc_d <- bf(response1M_sparc         ~ 0 + condition  *  difficulty_s + (1 + difficulty_s| condition:participant) + (1|condition:trial) )  

# Define priors
#prior0M <- get_prior( form0M_procustus , data = Data)
prior1M <- c(brms::prior(normal(0.5, .25), class = "b"))

# Run models
model_procustus1M <- run_model_cmdstanr(Data,form1M_procustus,"skew_normal",prior1M)
prior1M <- c(brms::prior(normal(0.5, .25), class = "b"))
model_sparc1M <- run_model_cmdstanr(Data,form1M_sparc,"skew_normal",prior1M)
prior1M_d <- c(brms::prior(normal(0.5, .25), class = "b"))
model_procustus1M_d <- run_model_cmdstanr(Data,form1M_procustus_d,"skew_normal",prior1M_d)
prior1M_d <- c(brms::prior(normal(10, .25), class = "b"))
model_sparc1M_d <- run_model_cmdstanr(Data,form1M_sparc_d,"skew_normal",prior1M_d)

# Save models
save(file = "Fit/model_procustus1M.RData", model_procustus1M)
save(file = "Fit/model_sparc1M.RData", model_sparc1M)
save(file = "Fit/model_procustus1M_d.RData", model_procustus1M_d)
save(file = "Fit/model_sparc1M_d.RData", model_sparc1M_d)
```

# non-calibrated and calibrated procustus Questionnaires WQP MPQS MPQP Difficulty
```{r questionnaires, echo=TRUE}
# Load data
load("Data.RData")

# Define priors
prior0 <- c(brms::prior(normal(0, 3), class = "b"))
prior0M <- c(brms::prior(normal(0, 3), class = "b"), 
             brms::prior(normal(0, 3), class = "b", coef = "log_response0M_procustus"))
prior1M <- c(brms::prior(normal(0, 3), class = "b"), 
             brms::prior(normal(0, 3), class = "b", coef = "response1M_procustus"))

# Define models for Questionnaires
formH3 <-  bf(WPQ ~ 0 + condition   + (1  |  condition:participant + condition:trial) )
formH3response0M_procustus <- bf(WPQ ~ 0 + log_response0M_procustus * condition   +  (1  |  condition:participant + condition:trial))
formH3response1M_procustus <- bf(WPQ ~ 0 + response1M_procustus * condition   +  (1  |  condition:participant + condition:trial) )

# Run models
modelH3_WPQ <- run_model_cmdstanr(Data,formH3,"gaussian",prior0)
modelH3response0M_WPQ <- run_model_cmdstanr(Data,formH3response0M_procustus,"gaussian",prior0M )
modelH3response1M_WPQ <- run_model_cmdstanr(Data,formH3response1M_procustus,"gaussian",prior1M )

# Define models for Questionnaires
formH3 <-  bf(MPQS ~ 0 + condition   + (1  |  condition:participant + condition:trial) )
formH3response0M_procustus <- bf(MPQS ~ 0 + log_response0M_procustus * condition +  (1  |  condition:participant + condition:trial))
formH3response1M_procustus <- bf(MPQS ~ 0 + response1M_procustus * condition +  (1  |  condition:participant + condition:trial) )

# Run models
modelH3_MPQS <- run_model_cmdstanr(Data,formH3,"gaussian",prior0)
modelH3response0M_MPQS <- run_model_cmdstanr(Data,formH3response0M_procustus,"gaussian",prior0M )
modelH3response1M_MPQS <- run_model_cmdstanr(Data,formH3response1M_procustus,"gaussian",prior1M )

# Define models for Questionnaires
formH3 <-  bf(MPQP ~ 0 + condition   + (1  |  condition:participant + condition:trial) )
formH3response0M_procustus <- bf(MPQP ~ 0 + log_response0M_procustus * condition  +   (1  |  condition:participant + condition:trial))
formH3response1M_procustus <- bf(MPQP ~ 0 + response1M_procustus * condition  +   (1  |  condition:participant + condition:trial) )

# Run models
modelH3_MPQP <- run_model_cmdstanr(Data,formH3,"gaussian",prior0)
modelH3response0M_MPQP <- run_model_cmdstanr(Data,formH3response0M_procustus,"gaussian",prior0M )
modelH3response1M_MPQP <- run_model_cmdstanr(Data,formH3response1M_procustus,"gaussian",prior1M )

# Define models for Questionnaires
formH3 <-  bf(Difficulty ~ 0 + condition + (1  |  condition:participant + condition:trial) )
formH3response0M_procustus <- bf(difficulty_s ~ 0 + log_response0M_procustus * condition  +  (1  |  condition:participant + condition:trial))
formH3response1M_procustus <- bf(difficulty_s ~ 0 + response1M_procustus * condition +  (1  |  condition:participant + condition:trial) )

# Run models
modelH3_Difficulty <- run_model_cmdstanr(Data,formH3,"gaussian",prior0)
modelH3response0M_Difficulty <- run_model_cmdstanr(Data,formH3response0M_procustus,"gaussian",prior0M )
modelH3response1M_Difficulty <- run_model_cmdstanr(Data,formH3response1M_procustus,"gaussian",prior1M )

# Save models
save(file = "Fit/modelH3_WPQ.RData",modelH3_WPQ)
save(file = "Fit/modelH3response0M_WPQ.RData",modelH3response0M_WPQ)
save(file = "Fit/modelH3response1M_WPQ.RData",modelH3response1M_WPQ)
save(file = "Fit/modelH3_MPQS.RData",modelH3_MPQS)
save(file = "Fit/modelH3response0M_MPQS.RData",modelH3response0M_MPQS)
save(file = "Fit/modelH3response1M_MPQS.RData",modelH3response1M_MPQS)
save(file = "Fit/modelH3_MPQP.RData",modelH3_MPQP)
save(file = "Fit/modelH3response0M_MPQP.RData",modelH3response0M_MPQP)
save(file = "Fit/modelH3response1M_MPQP.RData",modelH3response1M_MPQP)
save(file = "Fit/modelH3_Difficulty.RData",modelH3_Difficulty)
save(file = "Fit/modelH3response0M_Difficulty.RData",modelH3response0M_Difficulty)
save(file = "Fit/modelH3response1M_Difficulty.RData",modelH3response1M_Difficulty)
```
